requirements.txt
altgraph==0.17.4
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pywin32-ctypes==0.2.3
rawpy==0.24.0
setuptools==80.8.0
Werkzeug==3.1.3
yt-dlp==2025.3.31
